+++ CHANGELOG.md
@@ -86,6 +86,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 - Fixed [#1001](https://github.com/JabRef/jabref/issues/1001): No NPE when exporting a complete database
 - Fixed [#991](https://github.com/JabRef/jabref/issues/991): Entry is now correctly removed from the BibDatabase
 - Fixed [#1062](https://github.com/JabRef/jabref/issues/1062): Merge entry with DOI information now also applies changes to entry type
+- Fixed [#535](https://github.com/JabRef/jabref/issues/535): Add merge action to right click menu
 
 
 ### Removed